reload/requirements.txt
torch
torchvision
Pillow
gradio
pyserial
